Add GameState

+++ include/core/glfw.hpp
@@ -5,7 +5,7 @@
 
 namespace ntf::shogle {
 
-GLFWwindow* glfw_init(unsigned int w_width, unsigned int w_height, const char* w_name);
+GLFWwindow* glfw_init(size_t w_width, size_t w_height, const char* w_name);
 void glfw_destroy(GLFWwindow* win);
 
 }
